Eric Sunshine eliminated a number of compilation warnings throughout the project.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@20411 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/mesh/genmesh/persist/standard/gmeshldr.cpp
@@ -170,7 +170,7 @@ bool csGeneralFactoryLoader::ParseRenderBuffer(iDocumentNode *node,
   if (!buf.IsValid()) return false;
 
   size_t rbElem = buf->GetElementCount();
-  if (state->GetVertexCount() != rbElem)
+  if ((size_t)state->GetVertexCount() != rbElem)
   {
     synldr->ReportError ("crystalspace.genmeshfactoryloader.parse",
       node, "Render buffer vertex count(%u) different from "
@@ -686,7 +686,7 @@ bool csGeneralMeshLoader::ParseRenderBuffer(iDocumentNode *node,
   if (!buf.IsValid()) return false;
 
   size_t rbElem = buf->GetElementCount();
-  if (factstate->GetVertexCount() != rbElem)
+  if ((size_t)factstate->GetVertexCount() != rbElem)
   {
     synldr->ReportError ("crystalspace.genmeshloader.parse",
       node, "Render buffer vertex count(%u) different from "
